--- frontend/tests/unit/VSelect.spec.js
@@ -1,38 +1,0 @@
-//
-// Copyright (c) 2019 by SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-import { expect } from 'chai'
-import { shallowMount } from '@vue/test-utils'
-import Vue from 'vue'
-import Vuetify from 'vuetify'
-import { VSelect } from 'vuetify/lib'
-
-Vue.use(Vuetify)
-
-describe('VSelect', function () {
-  it('should be able to overwrite hint color class', function () {
-    const hint = 'hint test'
-    const propsData = {
-      hint,
-      'persistent-hint': true
-    }
-    const wrapper = shallowMount(VSelect, {
-      propsData
-    })
-    const hintElement = wrapper.find('.v-messages__wrapper > .v-messages__message')
-    expect(hintElement.text()).to.equal(hint)
-  })
-})
